LayoutTests/inspector/canvas/memory-expected.txt

Test that CanvasManager tracks canvas memory costs and is notified of changes.


== Running test suite: Canvas.memory
-- Running test case: Canvas.memory.memoryCost
Memory cost of canvas is NaN.

-- Running test case: Canvas.memory.canvasMemoryChanged
Change size of canvas to 200x200.
PASS: Memory cost should increase upon initially resizing canvas.
